Use double for all scaling values and float-point constant values at the Device Op API (#557)
* Use double as alpha/beta values type in reduce device op api

* Use double as alpha/beta values type in softmax device op api

* Use double as alpha/beta values type in multiple-reduce device op api

* Use double as epsilon value type in normalization/elementwise-normalization device op api

[ROCm/composable_kernel commit: 52abc2f37112d49f85f31aa343a14bd92a83b07c]

+++ client_example/15_reduce/reduce_nhwc_c.cpp
@@ -61,8 +61,8 @@ int main(int argc, char* argv[])
     for(auto dim : reduce_dims)
         reduce_length *= in_lengths[dim];
 
-    float alpha{1.0f};
-    float beta{0.0f};
+    double alpha{1.0};
+    double beta{0.0};
 
     SimpleDeviceMem in(sizeof(InDataType) * num_in_elements);
     SimpleDeviceMem out(sizeof(OutDataType) * num_out_elements);
